+++ api/core/db/models/evidence.py
@@ -9,9 +9,8 @@ class BasicEvidenceModel(BaseModel):
 
     Attributes: #noqa
         name (str): Description of collection that is tracked (e.g. 'view_details', 'purchase', 'order_entry).
-        user_uid (str): Unique identifier for user object.
+        user_uid (str, optional): Unique identifier for user object.
         item_id (str, optional): Can be used when tracking collection for interaction with items.
-        device_info (dict, optional): Capture devices infos to detect mobile/tablet/desktop views.
         path (str, optional): Specific URL path.
         timestamp (datetime): Current timestamp.
     """
@@ -19,7 +18,6 @@ class BasicEvidenceModel(BaseModel):
     name: str = Field(...)
     user_uid: Optional[str] = Field()
     item_id: Optional[str]
-    device_info: Optional[dict]
     path: Optional[str]
     timestamp: datetime = Field(default_factory=datetime.utcnow)
 
@@ -34,31 +32,6 @@ class BasicEvidenceModel(BaseModel):
                 "name": "view details",
                 "user": "123456789",
                 "item_id": "item1234",
-                "device_info": {
-                    "width": 500,
-                    "height": 700,
-                    "platform": "Win32",
-                    "language": "en-EN",
-                },
                 "path": "https://path-to-step.com"
             }
         }
-
-# TODO: user_id is not well defined (e.g. there needs to be a clearer definition)
-# -> maybe user_id comes from recommender system (for internal use only!)
-# -> collection 'user' has documents with _id=user_id (reco-system id!)
-# -> each user doc, contains a custom dictionary with other collected IDs
-# (e.g. "Business Partner ID", "device ID(s)", "reco cookie ID(s)")
-# but this means, an collection from an anonymous user (more general a user with a "probabilistic" user_id) needs to
-# be updated (user_id needs to be updated) as soon as the true (=deterministic) user_id reveals
-# Idea: what about a prob_user_id?
-# Idea: this means user_id can not be delivered by frontend or reco-interactor-js, reco-slider-js etc.
-# Idea: collection contains user_id but reco systems has user_uid
-# should I run an updater on collection collection to add user_uid to collection or should I create a link only
-# between user <> collection collection?
-# How do I store different user_id(s)? -> key-value pair? What is the key? Better key-array[values] pairs?
-# this means frontend identifier need to track an fixed attribute "identifier" which contains a dict, e.g.
-# identifier: {device-id: [123456, 654321], gp-id: abcde}
-# better: consumer_id (consumer coll. _id) is the reco-based user_id?
-# BEST: frontend: user_id, backend: consumer_id, user_id is contained in a list of identifiers of consumer document
-# collection can only contain user_id, backend always deals with consumer id?