+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Create attribute `created_time` for `BasicItemModel` and use for latest items route
 - Create delete by user uid for evidence
 - Get list of `BasicEvidenceModel` for `user_uid`
+- Store device info at user and not at evidence
 
 ## Version 0.2